Add data enable registers to existing devices
- Each device that generates data frames can be disabled to prevent
unused data from using bandwidth and power

+++ Bonsai.ONIX/BreakoutDigitalInputDataFrame.cs
@@ -1,21 +1,18 @@
 ﻿namespace Bonsai.ONIX
 {
-    /// <summary>
-    /// Provides Bonsai-friendly version of an AD7617DataBlock
-    /// </summary>
     public class BreakoutDigitalInputDataFrame : DataFrame
     {
         public BreakoutDigitalInputDataFrame(oni.Frame frame)
             : base(frame)
         {
             Port = sample[4];
-            Buttons = (ushort)(0x00FF & sample[5]);
-            Links = (ushort)((0x0F00 & sample[5]) >> 8);
+            Buttons = (byte)(0x00FF & sample[5]);
+            Links = (byte)((0x0F00 & sample[5]) >> 8);
         }
 
-        public ushort Buttons { get; private set; }
+        public byte Buttons { get; private set; }
 
-        public ushort Links { get; private set; }
+        public byte Links { get; private set; }
 
         public ushort Port { get; private set; }
